Egreso y UTI

+++ app/Http/Controllers/SalaController.php
@@ -42,10 +42,13 @@ class SalaController extends Controller
     public function eliminarsala($id){
         $salaEliminar=App\Models\Sala::findOrFail($id);
         $camas=App\Models\Cama::where('sala_id', '=', $id)->get();
+        if (($salaEliminar->id == 1) || ($salaEliminar->id == 2) ||($salaEliminar->id == 3)) {
+            return back()->with('mensaje2','No se puede borrar esta sala');
+        }
         foreach ($camas as $cama) {
             if ($cama->paciente_id != NULL) {
                 return back()->with('mensaje2','No se puede borrar una sala que tiene camas ocupadas');
-             }
+            }
         }
         $salaEliminar->delete();
         return back()->with('mensaje','Sala eliminada');
